# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenLaw API",
    "description": "REST API for OpenLaw. Also the third-party integration surface (TECH-003).",
    "version": "0.0.1",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    }
  },
  "components": {
    "schemas": {
...
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "default": "about:blank",
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": ["path", "message"],
              "additionalProperties": false
            }
          }
        },
        "required": ["type", "title", "status"],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/api/v1/contract-types/{id}/fields": {
      "get": {
        "operationId": "listContractTypeFields",
        "summary": "One contract type's attached fields in per-type order — the type editor's Attached fields card",
        "tags": ["contract-types"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "attachedFields": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "fieldId": {
                            "type": "string"
                          },
                          "slug": {
                            "type": "string"
                          },
                          "displayName": {
                            "type": "string"
                          },
                          "fieldType": {
                            "type": "string",
                            "enum": [
                              "text",
                              "long_text",
                              "number",
                              "date",
                              "boolean",
                              "single_select",
                              "multi_select",
                              "user",
                              "entity"
                            ]
                          },
                          "moduleScope": {
                            "type": "string",
                            "enum": ["contract", "global"]
                          },
                          "displayOrder": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "isRequired": {
                            "type": "boolean"
                          }
                        },
                        "required": [
                          "fieldId",
                          "slug",
                          "displayName",
                          "fieldType",
                          "moduleScope",
                          "displayOrder",
                          "isRequired"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": ["attachedFields"],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Problem details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "attachContractTypeField",
        "summary": "Attach a catalog field to a contract type: contract-scoped and global fields only (CTR-016), appended to the per-type order, optional from the start unless isRequired says otherwise",
        "tags": ["contract-types"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fieldId": {
                    "type": "string"
                  },
                  "isRequired": {
                    "type": "boolean"
                  }
                },
                "required": ["fieldId"]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "attachedField": {
                      "type": "object",
                      "properties": {
                        "fieldId": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "fieldType": {
                          "type": "string",
                          "enum": [
                            "text",
                            "long_text",
                            "number",
                            "date",
                            "boolean",
                            "single_select",
                            "multi_select",
                            "user",
                            "entity"
                          ]
                        },
                        "moduleScope": {
                          "type": "string",
                          "enum": ["contract", "global"]
                        },
                        "displayOrder": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "isRequired": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "fieldId",
                        "slug",
                        "displayName",
                        "fieldType",
                        "moduleScope",
                        "displayOrder",
                        "isRequired"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": ["attachedField"],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Problem details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}